+++ R/plot_missing.r
@@ -1,29 +1,37 @@
 #' Plot missing value profile
 #'
 #' This function returns and plots frequency of missing values for each feature.
-#' @param data input data to be profiled, in either \link{data.frame} or \link{data.table} format.
+#' @param data input data
 #' @param title plot title
+#' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
+#' @param theme_config a list of configurations to be passed to \link{theme}.
 #' @keywords plot_missing
 #' @aliases PlotMissing
 #' @details The returned object is suppressed by \link{invisible}.
+#' @details To change default font family and size, you may pass \code{base_size} and \code{base_family} to \code{ggtheme} options, e.g., \code{ggtheme = theme_gray(base_size = 15, base_family = "serif")}
+#' @details \code{theme_config} argument expects all inputs to be wrapped in a list object, e.g., to change the text color: \code{theme_config = list("text" = element_text(color = "blue"))}
 #' @return missing value information, such as frequency, percentage and suggested action.
 #' @import data.table
 #' @import ggplot2
 #' @importFrom scales comma
 #' @export plot_missing PlotMissing
 #' @examples
-#' # load packages
+#' # Load packages
 #' library(data.table)
 #'
-#' # manipulate data
+#' # Add missing values to iris data
 #' dt <- data.table(iris)
 #' for (j in 1:4) set(dt, i = sample(150, j * 30), j, value = NA_integer_)
 #'
-#' # plot and assign missing value information
+#' # Plot and assign missing value information
 #' na_profile <- plot_missing(dt)
 #' na_profile
+#'
+#' # Drop columns with more than 50% missing values
+#' drop_columns(dt, as.character(na_profile[pct_missing >= 0.5][["feature"]]))
+#' plot_missing(dt)
 
-plot_missing <- function(data, title = NULL) {
+plot_missing <- function(data, title = NULL, ggtheme = theme_gray(), theme_config = list("legend.position" = c("bottom"))) {
   ## Declare variable first to pass R CMD check
   feature <- num_missing <- pct_missing <- group <- NULL
   ## Check if input is data.table
@@ -31,9 +39,12 @@ plot_missing <- function(data, title = NULL) {
   ## Detect input data class
   data_class <- class(data)
   ## Set data to data.table
-  if (!is_data_table) {data <- data.table(data)}
+  if (!is_data_table) data <- data.table(data)
   ## Extract missing value distribution
-  missing_value <- data.table("feature" = names(data), "num_missing" = sapply(data, function(x) {sum(is.na(x))}))
+  missing_value <- data.table(
+    "feature" = names(data),
+    "num_missing" = sapply(data, function(x) {sum(is.na(x))})
+  )
   missing_value[, feature := factor(feature, levels = feature[order(-rank(num_missing))])]
   missing_value[, pct_missing := num_missing / nrow(data)]
   missing_value[pct_missing < 0.05, group := "Good"]
@@ -41,15 +52,18 @@ plot_missing <- function(data, title = NULL) {
   missing_value[pct_missing >= 0.4 & pct_missing < 0.8, group := "Bad"]
   missing_value[pct_missing >= 0.8, group := "Remove"][]
   ## Set data class back to original
-  if (!is_data_table) {class(missing_value) <- data_class}
+  if (!is_data_table) class(missing_value) <- data_class
   ## Create ggplot object
   output <- ggplot(missing_value, aes_string(x = "feature", y = "num_missing", fill = "group")) +
-    geom_bar(stat = "identity", colour = "black", alpha = 0.4) +
-    geom_text(aes(label = paste0(round(100 * pct_missing, 0), "%")), hjust = -0.15, size = 3.5) +
+    geom_bar(stat = "identity") +
+    geom_text(aes(label = paste0(round(100 * pct_missing, 2), "%"))) +
     scale_fill_manual("Group", values = c("Good" = "#1a9641", "OK" = "#a6d96a", "Bad" = "#fdae61", "Remove" = "#d7191c"), breaks = c("Good", "OK", "Bad", "Remove")) +
     scale_y_continuous(labels = comma) +
-    theme(legend.position = c("bottom")) + coord_flip() +
-    xlab("Features") + ylab("Number of missing rows") + ggtitle(title)
+    coord_flip() +
+    xlab("Features") + ylab("Number of missing rows") +
+    ggtitle(title) +
+    ggtheme +
+    do.call(theme, theme_config)
   ## Print plot
   print(output)
   ## Set return object