+++ R/plot_str.r
@@ -15,7 +15,6 @@
 #' @export plot_str PlotStr
 #' @seealso \link{str}
 #' @examples
-#' \dontrun{
 #' ## Visualize structure of iris dataset
 #' plot_str(iris)
 #'
@@ -33,19 +32,30 @@
 #' plot_str(obj, type = "d")
 #' ## Visualize only top 2 nested levels
 #' plot_str(obj, type = "d", max_level = 2)
-#' }
 
-plot_str <- function(data, type = c("diagonal", "radial"), max_level, print_network = TRUE, ...) {
+plot_str <- function(data, type = c("diagonal", "radial"), max_level = NULL, print_network = TRUE, ...) {
   ## Declare variable first to pass R CMD check
   i <- idx <- parent <- NULL
   ## Capture str output
-  str_output <- capture.output(str(data, vec.len = 0, give.attr = FALSE, give.length = FALSE))
+  str_output_raw <- capture.output(str(data, vec.len = 0, give.attr = FALSE, give.length = FALSE))
+  str_output <- unlist(lapply(str_output_raw, function(x) {gsub("\ \\.{2}\\@", "\\$\\@", x)}))
   n <- length(str_output)
   ## Split to calculate nested levels
   base_split <- tstrsplit(str_output[2:n], "\\$")
   nest_level <- (nchar(base_split[[1]]) - nchar(gsub("\ \\.{2}", "", base_split[[1]]))) / 3 + 1
+  ## Detect S4 objects
+  diff_nl <- diff(nest_level)
+  s4_start_index <- which(diff_nl > 1L) + 1L
+  if (length(s4_start_index) > 0) {
+    s4_end_index <- which(diff_nl == -2L)
+    s4_index_range <- unique(unlist(lapply(
+      s4_start_index,
+      function (i) {seq.int(i, s4_end_index[which.min(abs(s4_end_index - i))])}
+    )))
+    nest_level[s4_index_range] <- nest_level[s4_index_range] - 1L
+  }
   ## Handle max_level if exists
-  if (missing(max_level)) {
+  if (is.null(max_level)) {
     max_level <- max(nest_level)
   } else if (max_level <= 0 | max_level > max(nest_level)) {
     stop(paste0("max_level should be between 1 and ", max(nest_level)))
@@ -57,6 +67,7 @@ plot_str <- function(data, type = c("diagonal", "radial"), max_level, print_netw
   ## Make sure the root of each component has a unique name
   comp_root <- gsub(" ", "", comp_split[[1]])
   comp_root[which(comp_root == "")] <- make.names(comp_root[which(comp_root == "")], unique = TRUE)
+  if (anyDuplicated(comp_root)) comp_root[which(duplicated(comp_root))] <- paste0(comp_root[which(duplicated(comp_root))], "(name duplicated)")
   ## Combine component name with type
   comp_output <- paste0(comp_root, " (", trimws(gsub("NULL|\\.{3}|\\.{2}", "", comp_split[[2]])), ")")
   ## Transform data to table format
@@ -69,7 +80,7 @@ plot_str <- function(data, type = c("diagonal", "radial"), max_level, print_netw
   str_to_list <- function(str_dt, root_name = as.character(str_dt[["parent"]][1])) {
     str_list <- list(name = root_name)
     children <- str_dt[parent == root_name][["child"]]
-    if(length(children) > 0) {
+    if (length(children) > 0) {
       str_list[["children"]] <- lapply(children, str_to_list, str_dt = str_dt)
     }
     str_list