+++ R/split_columns.r
@@ -1,7 +1,7 @@
 #' Split data into discrete and continuous parts
 #'
 #' This function splits the input data into two \link{data.table} objects: discrete and continuous. A feature is continuous if \code{is.numeric} returns \code{TRUE}.
-#' @param data input data to be split, in either \link{data.frame} or \link{data.table} format.
+#' @param data input data
 #' @keywords split_columns
 #' @aliases SplitColType
 #' @details Features with all missing values will be dropped from the output data, but will be counted towards the column count.
@@ -27,10 +27,10 @@ split_columns <- function(data) {
   ## Detect input data class
   data_class <- class(data)
   ## Set data to data.table
-  if (!is_data_table) {data <- data.table(data)}
+  if (!is_data_table) data <- data.table(data)
   ## Find indicies for continuous features
-  ind <- sapply(data, is.numeric)
-  all_missing_ind <- sapply(data, function(x) {sum(is.na(x)) == length(x)})
+  all_missing_ind <- .getAllMissing(data)
+  ind <- sapply(data[, which(!all_missing_ind), with = FALSE], is.numeric)
   ## Count number of discrete, continuous and all-missing features
   n_all_missing <- sum(all_missing_ind)
   n_continuous <- sum(ind)
@@ -38,9 +38,9 @@ split_columns <- function(data) {
   ## Create object for continuous features
   continuous <- data[, which(ind), with = FALSE]
   ## Create object for discrete features
-  discrete <- data[, which(!(ind | all_missing_ind)), with = FALSE]
+  discrete <- data[, which(!ind), with = FALSE]
   ## Set data class back to original
-  if (!is_data_table) {class(discrete) <- class(continuous) <- data_class}
+  if (!is_data_table) class(discrete) <- class(continuous) <- data_class
   ## Set return object
   return(
     list(